backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "django>=5.1",
    "django-ninja>=1.3",
    "dj-database-url>=2.3",
    "gunicorn>=23.0",
    "psycopg[binary]>=3.2",
    "whitenoise>=6.8",
    "django-constance>=4.0",
    "markdown-it-py[linkify]>=3.0",
    "nh3>=0.2",
    "Pillow>=11.0",
    "pillow-heif>=1.2",
]

[dependency-groups]
dev = [
    "beautifulsoup4>=4.12",
    "pytest>=8.3",
    "pytest-django>=4.9",
    "requests>=2.32",
    "ruff>=0.8",
]
